Part 7 pt 2 Completed: edited volume to be double, rather than incorrect type: int

+++ hw-setup/src/main/java/setup/BallContainer.java
@@ -30,7 +30,7 @@ public class BallContainer implements Iterable<Ball> {
     /**
      * Volume of the BallContainer
      */
-    private int volume;
+    private double volume;
 
     /**
      * Constructor that creates a new BallContainer.
